moved some helper functions to more appropriate places
