+++ backend/ECondo.SharedKernel/Extensions/TypeExtensions.cs
@@ -1,0 +1,10 @@
+using ECondo.SharedKernel.Result;
+
+namespace ECondo.SharedKernel.Extensions;
+
+public static class TypeExtensions
+{
+    public static bool IsResultType(this Type type)
+        => type.IsGenericType &&
+           type.GetGenericTypeDefinition() == typeof(Result<,>);
+}
